requirements.txt
pandas
numpy
pgvector
nbformat
plotly
python-dotenv
SQLAlchemy
psycopg2
typing